move Options class to TCases command
