+++ tcases-lib/src/main/java/org/cornutum/tcases/Tcases.java
@@ -11,12 +11,9 @@ import org.cornutum.tcases.generator.*;
 import org.cornutum.tcases.generator.io.*;
 import org.cornutum.tcases.io.*;
 
-import java.io.File;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.util.HashMap;
 import java.util.Iterator;
-import java.util.Map;
 
 /**
  * Generates a set of {@link TestCase test cases} from a {@link SystemInputDef system input definition}.
@@ -24,272 +21,6 @@ import java.util.Map;
  */
 public class Tcases
   {
-  /**
-   * Represents a set of generator options.
-   */
-  public static class Options
-    {
-    public enum TransformType { HTML, JUNIT, CUSTOM };
-
-    /**
-     * Creates a new Options object.
-     */
-    public Options()
-      {
-      setExtended( true);
-      setWorkingDir( null);
-      }
-
-
-    /**
-     * Changes the output directory for generated test definitions.
-     */
-    public void setOutDir( File outDir)
-      {
-      outDir_ = outDir;
-      }
-
-    /**
-     * Returns the output directory for generated test definitions.
-     */
-    public File getOutDir()
-      {
-      return outDir_;
-      }
-
-    /**
-     * Changes the output file for generated test definitions.
-     */
-    public void setOutFile( File outFile)
-      {
-      outFile_ = outFile;
-      }
-
-    /**
-     * Returns the output file for generated test definitions.
-     */
-    public File getOutFile()
-      {
-      return outFile_;
-      }
-
-    /**
-     * Changes the output file for generated test definitions.
-     */
-    public void setTestDef( File testDef)
-      {
-      testDef_ = testDef;
-      }
-
-    /**
-     * Returns the output file for generated test definitions.
-     */
-    public File getTestDef()
-      {
-      return testDef_;
-      }
-
-    /**
-     * Changes the generator definition file.
-     */
-    public void setGenDef( File genDef)
-      {
-      genDef_ = genDef;
-      }
-
-    /**
-     * Returns the generator definition file.
-     */
-    public File getGenDef()
-      {
-      return genDef_;
-      }
-
-    /**
-     * Changes the transform file.
-     */
-    public void setTransformDef( File transformDef)
-      {
-      transformDef_ = transformDef;
-      }
-
-    /**
-     * Returns the transform file.
-     */
-    public File getTransformDef()
-      {
-      return transformDef_;
-      }
-
-    /**
-     * Changes the output transform type.
-     */
-    public void setTransformType( TransformType transformType)
-      {
-      transformType_ = transformType;
-      }
-
-    /**
-     * Returns the output transform type.
-     */
-    public TransformType getTransformType()
-      {
-      return transformType_;
-      }
-
-    /**
-     * Changes the transform parameter bindings.
-     */
-    public void setTransformParams( Map<String,Object> params)
-      {
-      transformParams_ = params;
-      }
-
-    /**
-     * Returns the transform parameter bindings.
-     */
-    public Map<String,Object> getTransformParams()
-      {
-      return transformParams_;
-      }
-
-    /**
-     * Changes the input definition file
-     */
-    public void setInputDef( File inputDef)
-      {
-      inputDef_ = inputDef;
-      }
-
-    /**
-     * Returns the input definition file
-     */
-    public File getInputDef()
-      {
-      return inputDef_;
-      }
-
-    /**
-     * Changes if new test definitions are generated by extending the
-     * previous {@link #getTestDef test definitions}.
-     */
-    public void setExtended( boolean extended)
-      {
-      extended_ = extended;
-      }
-
-    /**
-     * Returns if new test definitions are generated by extending the
-     * previous {@link #getTestDef test definitions}.
-     */
-    public boolean isExtended()
-      {
-      return extended_;
-      }
-
-    /**
-     * Changes the random seed used by generators.
-     */
-    public void setRandomSeed( Long seed)
-      {
-      seed_ = seed;
-      }
-
-    /**
-     * Returns the random seed used by generators.
-     */
-    public Long getRandomSeed()
-      {
-      if( seed_ == null && isNewSeed())
-        {
-        setRandomSeed( (long) (Math.random() * Long.MAX_VALUE));
-        }
-
-      return seed_;
-      }
-
-    /**
-     * Changes if choosing a new random seed used by generators.
-     */
-    public void setNewSeed( boolean newSeed)
-      {
-      newSeed_ = newSeed;
-      }
-
-    /**
-     * Returns if choosing a new random seed used by generators.
-     */
-    public boolean isNewSeed()
-      {
-      return newSeed_;
-      }
-
-    /**
-     * Changes the default tuple size used by generators.
-     */
-    public void setDefaultTupleSize( Integer tupleSize)
-      {
-      defaultTupleSize_ = tupleSize;
-      }
-
-    /**
-     * Returns the default tuple size used by generators.
-     */
-    public Integer getDefaultTupleSize()
-      {
-      return defaultTupleSize_;
-      }
-
-    /**
-     * Changes the current working directory used to complete relative path names.
-     */
-    public void setWorkingDir( File workingDir)
-      {
-      workingDir_ =
-        workingDir == null
-        ? new File( ".")
-        : workingDir;
-      }
-
-    /**
-     * Returns the current working directory used to complete relative path names.
-     */
-    public File getWorkingDir()
-      {
-      return workingDir_;
-      }
-
-    /**
-     * Changes if the current version should be shown.
-     */
-    public void setShowVersion( boolean showVersion)
-      {
-      showVersion_ = showVersion;
-      }
-
-    /**
-     * Returns if the current version should be shown.
-     */
-    public boolean showVersion()
-      {
-      return showVersion_;
-      }
-
-    private File inputDef_;
-    private File outDir_;
-    private File outFile_;
-    private File testDef_;
-    private File genDef_;
-    private File transformDef_;
-    private Map<String,Object> transformParams_ = new HashMap<String,Object>();
-    private TransformType transformType_;
-    private boolean extended_;
-    private Long seed_;
-    private boolean newSeed_;
-    private Integer defaultTupleSize_;
-    private File workingDir_;
-    private boolean showVersion_;
-    }
 
   /**
    * Creates a new Tcases object.
@@ -303,15 +34,12 @@ public class Tcases
    * base test definitions. If <CODE>genDef</CODE> is null, the default generator is used.
    * If <CODE>baseDef</CODE> is null, no base test definitions are used.
    */
-  public static SystemTestDef getTests( SystemInputDef inputDef, IGeneratorSet genDef, SystemTestDef baseDef, Options options)
+  public static SystemTestDef getTests( SystemInputDef inputDef, IGeneratorSet genDef, SystemTestDef baseDef, Long seed, Integer defaultTupleSize)
     {
     if( genDef == null)
       {
       genDef = GeneratorSet.basicGenerator();
       }
-
-    Long seed = options==null? null : options.getRandomSeed();
-    Integer defaultTupleSize = options==null? null : options.getDefaultTupleSize();
 
     SystemTestDef testDef = new SystemTestDef( inputDef.getName());
     for( Iterator<FunctionInputDef> functionDefs = inputDef.getFunctionInputDefs(); functionDefs.hasNext();)
@@ -349,7 +77,7 @@ public class Tcases
    */
   public static SystemTestDef getTests( SystemInputDef inputDef, IGeneratorSet genDef, SystemTestDef baseDef)
     {
-    return getTests( inputDef, genDef, baseDef, null);
+    return getTests( inputDef, genDef, baseDef, null, null);
     }
 
   /**
